Adjustments for better gameplay, bigger rect for collisions, fix obstacle_manager.py, reduced points bonus from near overtaking
